--- assets/css/style.css
@@ -1,80 +1,0 @@
-* {
- box-sizing: border-box;
- margin: 0;
- padding: 0;
-}
-
-body {
-  background-image: url(../img/pexels-photo-301673.jpeg);
-  width: 100%;
-  min-height: 100%;
-  background-repeat: no-repeat;
-}
-
-/* Div background blanco con opacidad */
-.white-opacity{
-  background-color: #FFFFFF; 
-  opacity: 0.7;
-  margin:2.7em;
-  border-radius: 3vh;
-  padding-bottom: 2em;
-  margin-bottom: 1em;
-}
-
-/* H1, título principal */
-.title1{
-  font-size: 9vh;
-  color: #FF5881;
-  text-align: center;
-  padding-bottom: 1.4vw;
-}
-
-/* Estilo para texto descriptivo */
-.description-en{
-  font-size: 1.6em;
-  margin-left: 2vw;
-  margin-right: 2vw;
-  text-align: center;
-}
-
-/* Subtítulos para divs */
-.subtitle{
-  color:#3E3E3E;
-  margin-top: 8vw;
-  margin-bottom: 8vw;
-}
-
-/* Div movimiento horizontal */
-
-.square{
-  width: 12em;
-  height: 12em;
-  background-color: #FF475A;
-  border: 2px solid #FF322F;
-  border-radius: 3px;
-  margin-bottom: 1em;
-}
-
-.left{
-  float:right;
-}
-
-/* Estilo Desplegable */
-
-.down{
-  /* border: 1px solid red Borrar esto después */
-}
-
-.plegable{
-  height:0;
-  overflow:hidden;
-}
-
-#toTheRight, #toTheLeft, #down {
-  height: 70vh;
-}
-
-.head {
-  margin-bottom: 100vh;
-
-}
