--- index.html
@@ -1,65 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta charset="UTF-8">
-    <title>Library Project</title>
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap-theme.min.css" integrity="sha384-rHyoN1iRsVXV4nD0JutlnGaslCJuC7uwjduW9SVrLvRYooPp2bWYgmgJQIXwl/Sp" crossorigin="anonymous">
-    <link rel="stylesheet" href="assets/css/style.css">
-    <script src="https://use.fontawesome.com/c983d8eba6.js"></script>
-  </head>
-  <body>
-    <section class="head white-opacity">
-      <div class="container-fluid">
-        <div class="row">
-          <div class="col-sm-12 col-md-12 col-lg-12 col-xl-12">
-            <h1 class="title1 center-block">A(ddi)ctive</h1>
-          </div>
-        </div>
-        <div class="row">
-          <div class="col-sm-12 col-md-12 col-lg-12 col-xl-12">
-            <p class="description-en">This is a mini-library you can use in your webpage to draw the attention of your users with fun animations given to any div (if not all!) of your html. Enjoy it!</p>
-          </div>
-        </div>
-      </div>
-    </section>
-    <section id="toTheRight">
-      <div class="container">
-        <div class="row">
-          <h2 class="text-center subtitle">This square moves to the right</h2>
-          <div id="divToRight" class="square right">
-            <!-- This div will move to the right and it will stop in the center of the webpage -->
-          </div>
-        </div>
-      </div>
-    </section>
-    <section id="toTheLeft">
-      <div class="container">
-        <div class="row">
-          <h2 class="text-center subtitle">This square moves to the left</h2>
-          <div id="divToLeft" class="square left">
-            <!-- This div will move to the left and it will stop in the center of the webpage -->
-          </div>
-        </div>
-      </div>
-    </section>
-    <section id="down" class="down">
-      <div class="container">
-        <div class="row">
-          <h2 class="text-center subtitle desple">This is a drop div... Put your mouse over!</h2>
-          <div class="plegable center-block">
-            <img src="https://media.giphy.com/media/mVJ5xyiYkC3Vm/giphy.gif" alt="">
-            <p>Holi, soy un desplegable.</p>
-          </div>
-        </div>
-      </div>
-    </section>
-    
-  </body>
-  <script src="node_modules/jquery/dist/jquery.min.js"></script>
-  <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
-  <script src="assets/js/app.js"></script>
-    <script src="node_modules/chai/chai.js"></script>
-  <script src="node_modules/chai-jquery/chai-jquery.js"></script>
-</html>
